requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Werkzeug==2.2.2
gunicorn==25.0.3
python-dotenv
git+https://github.com/MrPeterss/infosci-spark-client.git
scikit-learn==1.7.2
matplotlib==3.7.2
nltk
infosci-spark-client @ git+https://github.com/MrPeterss/infosci-spark-client.git@20b5e9dc8f9bb3286ee907aba7dd7f4fc4167d79